+++ glaze-lib/webview/webview-stub.rkt
@@ -19,6 +19,7 @@
                       #:title [title "Glaze"]
                       #:width [width 1024]
                       #:height [height 768]
+                      #:devtools? [devtools? #f]
                       #:on-close [on-close (lambda () (void))])
   #f)